Feat: 가입 프로필설정 이미지 해결
